% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bookman.R
\name{bookman_server}
\alias{bookman_server}
\title{dv.bookman module server.}
\usage{
bookman_server(id = "_bookman", bookmark_dir = NULL)
}
\arguments{
\item{id}{Unique shiny string ID. Should match the ID used in \code{bookman_UI()}.}

\item{bookmark_dir}{Path to the bookmark folder of the application. Autodetected if \code{NULL}}
}
\description{
dv.bookman module server.
}
