% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bookman.R
\name{bookman_UI}
\alias{bookman_UI}
\title{dv.bookman module UI.}
\usage{
bookman_UI(id = "_bookman")
}
\arguments{
\item{id}{Unique shiny string ID. Should match the ID used in \code{bookman_server()}.}
}
\description{
dv.bookman module UI.
}
