% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bookman.R
\name{mod_bookman}
\alias{mod_bookman}
\title{dv.bookman DaVinci module.}
\usage{
mod_bookman(module_id = "_bookman", bookmark_dir = NULL)
}
\arguments{
\item{module_id}{Unique shiny string ID.}

\item{bookmark_dir}{Path to the bookmark folder of the application. Autodetected if \code{NULL}.}
}
\description{
See this vignette for more details: \code{vignette("manual", package = "dv.bookman")}
}
